backend/requirements.txt
python-telegram-bot==20.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.27
alembic==1.13.1
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
gunicorn==21.2.0
pandas==2.2.0
numpy==1.26.3 